move compute from tof to separate file, add new compute from chopper_cascade
