+++ packages/essreduce/src/ess/reduce/unwrap/types.py
@@ -14,6 +14,63 @@ LookupTableFilename = NewType("LookupTableFilename", str)
 """Filename of the wavelength lookup table."""
 
 
+LtotalRange = NewType("LtotalRange", tuple[sc.Variable, sc.Variable])
+"""
+Range (min, max) of the total length of the flight path from the source to the detector.
+This is used to create the lookup table to compute the neutron time-of-flight.
+Note that the resulting table will extend slightly beyond this range, as the supplied
+range is not necessarily a multiple of the distance resolution.
+
+Note also that the range of total flight paths is supplied manually to the workflow
+instead of being read from the input data, as it allows us to compute the expensive part
+of the workflow in advance (the lookup table) and does not need to be repeated for each
+run, or for new data coming in in the case of live data collection.
+"""
+
+DistanceResolution = NewType("DistanceResolution", sc.Variable)
+"""
+Step size of the distance axis in the lookup table.
+Should be a single scalar value with a unit of length.
+This is typically of the order of 1-10 cm.
+"""
+
+
+@dataclass
+class SourcePulse:
+    time: tuple[sc.Variable, sc.Variable]
+    """Time range (start, end) of the source pulse."""
+    wavelength: tuple[sc.Variable, sc.Variable]
+    """Wavelength range (min, max) of the neutrons in the source pulse."""
+
+
+TimeResolution = NewType("TimeResolution", sc.Variable)
+"""
+Step size of the event_time_offset axis in the lookup table.
+This is basically the 'time-of-flight' resolution of the detector.
+Should be a single scalar value with a unit of time.
+This is typically of the order of 0.1-0.5 ms.
+
+Since the event_time_offset range needs to span exactly one pulse period, the final
+resolution in the lookup table will be at least the supplied value here, but may be
+smaller if the pulse period is not an integer multiple of the time resolution.
+"""
+
+PulsePeriod = NewType("PulsePeriod", sc.Variable)
+"""
+Period of the source pulses, i.e., time between consecutive pulse starts.
+"""
+
+PulseStride = NewType("PulseStride", int)
+"""
+Stride of used pulses. Usually 1, but may be a small integer when pulse-skipping.
+"""
+
+SourcePosition = NewType("SourcePosition", sc.Variable)
+"""
+Position of the neutron source in the coordinate system of the choppers.
+"""
+
+
 @dataclass
 class LookupTable:
     """
